% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_z_score_power.R
\name{average_z_score_power}
\alias{average_z_score_power}
\title{Generate simulated data and calculate power of average z-score approach.}
\usage{
average_z_score_power(
  num_sims,
  num_observations,
  num_endpoints,
  type_vector,
  mean_vector,
  sd_vector,
  effect_vector,
  cor_mat = NA,
  direction_vector,
  threshold = 0.05,
  start.seed = 123456
)
}
\arguments{
\item{num_sims}{A numeric value representing the number of simulations to be run.}

\item{num_observations}{A numeric value representing the number of observations
per arm. Currently only two-arms with equal allocation is available.}

\item{num_endpoints}{A numeric value between 1 and 5 that represents the number
of endpoints being used. It is not recommended to have more than 5 endpoints.}

\item{type_vector}{A character vector of length \code{num_endpoints} that specifies
the type of endpoints. Currently normal (\code{n}), log-normal (\code{l}),
and binary (\code{b}) are supported.}

\item{mean_vector}{A numeric vector of length \code{num_endpoints} that specifies
the expected mean values in the control arm. Log-normal means represent the
mean of follow/baseline values; binary means represent the proportion of
events or successes.}

\item{sd_vector}{A numeric vector of length \code{num_endpoints} that specifies
the expected standard deviation (sd) values in the control arm.
Log-normal sds represent the sd of log(follow/baseline) values.
Binary sds are calculated within the function and should be listed as \code{NA}.}

\item{effect_vector}{A numeric vector of length \code{num_endpoints} that specifies
the expected relative treatment effect for each endpoint (i.e., 1.2 represents
a 20\% increase in the treatment arm).}

\item{cor_mat}{A numeric matrix of dimensions \code{num_endpoints} x \code{num_endpoints}
that specifies the correlations between all endpoints. The diagnonal of the
correlation matrix should equal all 1s. If not specified, no correlations
between any endpoints will be assumed. \bold{(optional)}}

\item{direction_vector}{A numeric vector of length \code{num_endpoints} that specifies
the direction of positive treatment effect. All elements should be either a
1 (larger values are better), or a -1 (smaller values are better).}

\item{threshold}{A numeric value that specifies the p-value cutoff for
statistical significance, based on a two-sided test. If not specified, a
threshold of 0.05 will be assumed. \bold{(optional)}}

\item{start.seed}{A numeric value that specifies the starting seed for
simulation of the multivariate normal responses. \bold{(optional)}}
}
\description{
\code{average_z_score_power} calculates simulated power for two-arm average z-score approach.
}
\examples{
test_scores <- average_z_score_power(num_sims = 100,
num_observations = 100,
num_endpoints = 5,
type_vector = c("n", "n", "l", "n", "b"),
mean_vector = c(100, -5, 0.8, 30, 0.3),
sd_vector = c(75, 5, 0.8, 20, NA),
effect_vector = c(1.2, 1.2, 0.8, 1.2, 0.8),
cor_mat = matrix(c(1, -0.3, -0.3, 0, -0.5,
                   -0.3, 1, 0, 0, 0,
                   -0.3, 0, 1, 0, 0.4,
                   0, 0, 0, 1, 0,
                   -0.5, 0, 0.4, 0, 1),
                 nrow = 5, byrow = T),
direction_vector = c(1, -1, -1, 1, -1),
threshold = 0.05,
start.seed = 123456)

test_scores$power


}
